feat: implement performance benchmarking methods and reporting in PerformanceBenchmark class
