+++ Sources/SpeechTranscriber+LanguageDetection.swift
