+++ Sources/SpeechTranscriber+SegmentProcessing.swift
